+++ examples/store-reactjs-chat/src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import { getStatusFleetNodes, StoreCodec, Waku } from 'js-waku';
+import { getBootstrapNodes, StoreCodec, Waku } from 'js-waku';
 import * as React from 'react';
 import protons from 'protons';
 
@@ -101,8 +101,12 @@ function App() {
 export default App;
 
 async function bootstrapWaku(waku) {
-  const nodes = await getStatusFleetNodes();
-  await Promise.all(nodes.map((addr) => waku.dial(addr)));
+  try {
+    const nodes = await getBootstrapNodes();
+    await Promise.all(nodes.map((addr) => waku.dial(addr)));
+  } catch (e) {
+    console.error('Failed to bootstrap to Waku network');
+  }
 }
 
 function decodeMessage(wakuMessage) {