+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Renamed `discover.getStatusFleetNodes` to `discovery.getBootstrapNodes` and make it more generic to allow retrieval of bootstrap nodes from other sources.
+
 ### Removed
 - Examples (cli-chat): The focus of this library is Web environment;
   Several examples now cover usage of Waku Relay and Waku Store making cli-chat example obsolete;